Load the machineIds onTraceLoad instead of LogPanel constructor
